+++ packages/weapp-tailwindcss/src/bundlers/gulp/shared/create-native-framework-plugins.ts
@@ -1,7 +1,6 @@
 import type File from 'vinyl'
 import type { BundleRuntimeClassSetManager } from '../../vite/incremental-runtime-class-set'
-import type { CreateJsHandlerOptions, IStyleHandlerOptions, ITemplateHandlerOptions, JsModuleGraphOptions, UserDefinedOptions } from '@/types'
-import { Buffer } from 'node:buffer'
+import type { IStyleHandlerOptions, ITemplateHandlerOptions, JsModuleGraphOptions, UserDefinedOptions } from '@/types'
 import path from 'node:path'
 import process from 'node:process'
 import { prependConfigDirective } from '@/bundlers/shared/generator-css/config-directive'
@@ -9,113 +8,23 @@ import { hasTailwindRootDirectives, normalizeTailwindConfigDirectives, normalize
 import { getCompilerContext } from '@/context'
 import { normalizeStyleHandlerMajorVersion } from '@/context/style-options'
 import { createDebug } from '@/debug'
-import { shouldSkipJsTransform } from '@/js/precheck'
 import { createTailwindRuntimeReadyPromise, ensureRuntimeClassSet } from '@/tailwindcss/runtime'
 import { getRuntimeClassSetSignature } from '@/tailwindcss/runtime/cache'
-import { getTailwindV4IncrementalGenerateCacheStats } from '@/tailwindcss/v4-engine'
 import { hasConfiguredTailwindV4CssRoots, upsertTailwindV4CssSource } from '@/tailwindcss/v4/css-sources'
-import { processCachedTask } from '../../shared/cache'
-import { annotateCssSourceTrace, createCssSourceTraceCacheSignature, createCssTokenSourceMap } from '../../shared/css-source-trace'
-import { createBundlerGeneratedCssMarker, hasBundlerGeneratedCssMarker, stripBundlerGeneratedCssMarkers } from '../../shared/generated-css-marker'
-import { generateCssByGenerator } from '../../shared/generator-css'
-import { finalizeMiniProgramGeneratorCss } from '../../shared/generator-css/generation-helpers'
-import { rewriteLocalCssImportRequestsForOutput, splitLocalCssImports } from '../../shared/generator-css/local-imports'
+import { splitLocalCssImports } from '../../shared/generator-css/local-imports'
 import { createBundleRuntimeClassSetManager } from '../../vite/incremental-runtime-class-set'
 import { createSourceCandidateCollector } from '../../vite/source-candidates'
 import { resolveViteSourceScanEntries } from '../../vite/source-scan'
 import { createGulpModuleGraphOptions } from '../module-graph'
 import { createGulpRuntimeSnapshot } from '../runtime-snapshot'
-import { createVinylTransform } from '../vinyl-transform'
+import {
+  pruneGulpRuntimeSourceCaches,
+  resolveGulpMemoryDebugStats,
+  touchMapEntry,
+} from './create-native-framework-plugins/cache-state'
+import { createGulpFileTransforms } from './create-native-framework-plugins/file-transforms'
 
 const debug = createDebug()
-const GULP_RUNTIME_SOURCE_CACHE_MAX = 256
-const GULP_PROCESS_CACHE_MAX = 512
-
-function toMb(bytes: number) {
-  return Math.round(bytes / 1024 / 1024)
-}
-
-function touchMapEntry<Key, Value>(map: Map<Key, Value>, key: Key, value: Value) {
-  map.delete(key)
-  map.set(key, value)
-}
-
-function pruneGulpRuntimeSourceCaches(
-  sourceHashByFile: Map<string, string>,
-  sourcesByFile: Map<string, { source: string, type: 'html' | 'js' }>,
-) {
-  while (sourcesByFile.size > GULP_RUNTIME_SOURCE_CACHE_MAX) {
-    const oldestKey = sourcesByFile.keys().next().value
-    if (typeof oldestKey !== 'string') {
-      break
-    }
-    sourcesByFile.delete(oldestKey)
-    sourceHashByFile.delete(oldestKey)
-  }
-}
-
-function rememberGulpProcessCacheKey(cacheKeys: Set<string>, key: string) {
-  cacheKeys.delete(key)
-  cacheKeys.add(key)
-  while (cacheKeys.size > GULP_PROCESS_CACHE_MAX) {
-    const oldestKey = cacheKeys.keys().next().value
-    if (typeof oldestKey !== 'string') {
-      break
-    }
-    cacheKeys.delete(oldestKey)
-  }
-}
-
-function pruneGulpProcessCache(cache: ReturnType<typeof getCompilerContext>['cache'], cacheKeys: Set<string>) {
-  cache.prune?.({
-    cacheKeys,
-    hashKeys: cacheKeys,
-  })
-}
-
-function resolveGulpMemoryDebugStats(context: {
-  cache: ReturnType<typeof getCompilerContext>['cache']
-  defaultStyleHandlerOptionsCache: Map<number | 'unknown', Partial<IStyleHandlerOptions>>
-  gulpProcessCacheKeys: Set<string>
-  phase: string
-  runtimeSet: Set<string>
-  runtimeSourceHashByFile: Map<string, string>
-  runtimeSourcesByFile: Map<string, { source: string, type: 'html' | 'js' }>
-}) {
-  if (process.env['WEAPP_TW_HMR_MEMORY_DEBUG'] !== '1') {
-    return undefined
-  }
-
-  const memory = process.memoryUsage()
-  return {
-    phase: context.phase,
-    process: {
-      rssMb: toMb(memory.rss),
-      heapTotalMb: toMb(memory.heapTotal),
-      heapUsedMb: toMb(memory.heapUsed),
-      externalMb: toMb(memory.external),
-      arrayBuffersMb: toMb(memory.arrayBuffers),
-    },
-    runtime: {
-      runtimeSet: context.runtimeSet.size,
-      runtimeSourceHashByFile: context.runtimeSourceHashByFile.size,
-      runtimeSourcesByFile: context.runtimeSourcesByFile.size,
-      maxRuntimeSources: GULP_RUNTIME_SOURCE_CACHE_MAX,
-    },
-    processCache: {
-      instance: context.cache.instance.size,
-      hashMap: context.cache.hashMap.size,
-      activeCacheKeys: context.gulpProcessCacheKeys.size,
-      maxCacheKeys: GULP_PROCESS_CACHE_MAX,
-    },
-    gulpOptions: {
-      defaultStyleHandlerOptions: context.defaultStyleHandlerOptionsCache.size,
-    },
-    tailwind: {
-      v4: getTailwindV4IncrementalGenerateCacheStats(),
-    },
-  }
-}
 
 /**
  * @name weapp-tw-gulp
@@ -129,7 +38,7 @@ export function createNativeGulpPlugins(options: UserDefinedOptions = {}) {
     __internalDeferMissingCssEntriesWarning: true,
   } as UserDefinedOptions)
 
-  const { templateHandler, styleHandler, jsHandler, cache } = opts
+  const { cache } = opts
   const initialTailwindRuntime = opts.tailwindRuntime
   const refreshTailwindcssRuntime = opts.refreshTailwindcssRuntime
 
@@ -429,223 +338,26 @@ export function createNativeGulpPlugins(options: UserDefinedOptions = {}) {
     }
   }
 
-  function createWxssTransform(
-    options: Partial<IStyleHandlerOptions>,
-    stage: 'generate' | 'transform',
-  ) {
-    return createVinylTransform(`css:${stage}`, async (file) => {
-      if (!file.contents) {
-        return
-      }
-      const rawSource = file.contents.toString()
-      const cssSourceChanged = await registerAutoCssSource(file, rawSource)
-      const shouldUseGenerator = runtimeState.tailwindRuntime.majorVersion === 4
-      let gulpSourceCandidateGetter: typeof cachedGulpSourceCandidateGetter
-      if (shouldUseGenerator) {
-        gulpSourceCandidateGetter = await refreshGulpV4SourceCandidates(cssSourceChanged)
-      }
-      const nextRuntimeSet = await refreshRuntimeSet({
-        forceRefresh: cssSourceChanged,
-        forceCollect: cssSourceChanged,
-        clearCache: cssSourceChanged,
-      })
-      const sourceTraceTokenSources = cachedGulpSourceCandidateSourceGetter
-        ? createCssTokenSourceMap(cachedGulpSourceCandidateSourceGetter(undefined), opts)
-        : undefined
-      const sourceTraceSignature = createCssSourceTraceCacheSignature(sourceTraceTokenSources, opts)
-      const sourceCandidateSignature = cachedGulpSourceCandidateGetter
-        ? `gulp-source-candidates:1:${[...cachedGulpSourceCandidateGetter(undefined)].sort().join('\n')}`
-        : undefined
-      const styleOutputExtension = resolveGulpStyleOutputExtension(file)
-      const outputSignature = styleOutputExtension
-        ? `gulp-output:1:${stage}:${styleOutputExtension}`
-        : `gulp-output:0:${stage}`
-      await processCachedTask<string>({
-        cache,
-        cacheKey: file.path,
-        hash: createRuntimeSetHash(rawSource, nextRuntimeSet, sourceTraceSignature, sourceCandidateSignature, outputSignature),
-        applyResult(source) {
-          file.contents = Buffer.from(source)
-        },
-        onCacheHit() {
-          debug('css cache hit: %s', file.path)
-        },
-        async transform() {
-          await runtimeState.readyPromise
-          const cssHandlerOptions = resolveWxssFileHandlerOptions(file, rawSource, options)
-          const generated = shouldUseGenerator
-            ? await generateCssByGenerator({
-                opts,
-                runtimeState,
-                runtime: nextRuntimeSet,
-                rawSource,
-                file: file.path,
-                cssHandlerOptions,
-                cssUserHandlerOptions: resolveWxssUserHandlerOptions(options),
-                getSourceCandidatesForEntries: gulpSourceCandidateGetter,
-                styleHandler,
-                debug,
-                deferCssAdaptation: stage === 'generate',
-              })
-            : undefined
-          const transformedCss = generated?.css ?? (stage === 'generate'
-            ? rawSource
-            : (await styleHandler(rawSource, cssHandlerOptions)).css)
-          if (stage === 'generate') {
-            const preflightMode = generated?.metadata?.preflightMode
-            if (preflightMode) {
-              generatedCssPreflightModeByFile.set(file.path, preflightMode)
-            }
-            else {
-              generatedCssPreflightModeByFile.delete(file.path)
-            }
-          }
-          const css = annotateCssSourceTrace(transformedCss, {
-            opts,
-            tokenSources: sourceTraceTokenSources,
-          })
-          const generatedCss = stage === 'generate' && generated
-            ? `${createBundlerGeneratedCssMarker('gulp', file.path)}\n${css}`
-            : css
-          const outputCss = stage === 'generate'
-            ? generatedCss
-            : rewriteLocalCssImportRequestsForOutput(generatedCss, {
-                styleOutputExtension,
-              })
-          debug('css handle: %s', file.path)
-          return {
-            result: outputCss,
-          }
-        },
-      })
-      rememberGulpProcessCacheKey(gulpProcessCacheKeys, file.path)
-      pruneGulpProcessCache(cache, gulpProcessCacheKeys)
-    }, () => resolveGulpTransformTimingDetails(`css:${stage}`))
-  }
-
-  // 显式生成阶段保留源码 import，交给后续 gulp-postcss 解析。
-  const generateWxss = (options: Partial<IStyleHandlerOptions> = {}) => createWxssTransform(options, 'generate')
-  const transformWxss = (options: Partial<IStyleHandlerOptions> = {}) => createWxssTransform(options, 'transform')
-  const adaptWxss = (options: Partial<IStyleHandlerOptions> = {}) =>
-    createVinylTransform('css:adapt', async (file) => {
-      if (!file.contents) {
-        return
-      }
-      const rawSource = file.contents.toString()
-      const generatedCss = hasBundlerGeneratedCssMarker(rawSource)
-      const source = generatedCss ? stripBundlerGeneratedCssMarkers(rawSource) : rawSource
-      const styleOutputExtension = resolveGulpStyleOutputExtension(file)
-      const cssHandlerOptions = resolveWxssFileHandlerOptions(file, source, options)
-      const handled = await styleHandler(source, cssHandlerOptions)
-      const preflightMode = generatedCssPreflightModeByFile.get(file.path)
-      const finalized = generatedCss
-        ? finalizeMiniProgramGeneratorCss(
-            handled.css,
-            'weapp',
-            runtimeState.tailwindRuntime.majorVersion,
-            opts.cssPreflight,
-            {
-              injectPreflight: preflightMode?.inject ?? cssHandlerOptions.isMainChunk,
-              preservePreflight: preflightMode?.preserve ?? cssHandlerOptions.isMainChunk,
-              styleOptions: cssHandlerOptions,
-            },
-          )
-        : handled.css
-      file.contents = Buffer.from(rewriteLocalCssImportRequestsForOutput(finalized, {
-        styleOutputExtension,
-      }))
-      debug('css adapt: %s', file.path)
-    }, () => resolveGulpTransformTimingDetails('css:adapt'))
-
-  const transformJs = (options: Partial<CreateJsHandlerOptions> = {}) =>
-    createVinylTransform('js', async (file) => {
-      if (!file.contents) {
-        return
-      }
-      const filename = path.resolve(file.path)
-      const rawSource = file.contents.toString()
-      await refreshRuntimeSetForSource(file, rawSource, 'js')
-      await runtimeState.readyPromise
-      const moduleGraph = resolveModuleGraphOptions(options.moduleGraph)
-      const handlerOptions: CreateJsHandlerOptions = {
-        ...options,
-        generateMap: false,
-        filename,
-        moduleGraph,
-        babelParserOptions: {
-          ...(options?.babelParserOptions ?? {}),
-          sourceFilename: filename,
-        },
-      }
-      if (runtimeState.tailwindRuntime.majorVersion !== undefined) {
-        handlerOptions.tailwindcssMajorVersion = runtimeState.tailwindRuntime.majorVersion
-      }
-      await processCachedTask<string>({
-        cache,
-        cacheKey: file.path,
-        rawSource,
-        applyResult(source) {
-          file.contents = Buffer.from(source)
-        },
-        onCacheHit() {
-          debug('js cache hit: %s', file.path)
-        },
-        async transform() {
-          await runtimeState.readyPromise
-          const currentSource = file.contents?.toString() ?? rawSource
-          if (shouldSkipJsTransform(currentSource, {
-            ...handlerOptions,
-            classNameSet: runtimeSet,
-          } as CreateJsHandlerOptions)) {
-            return { result: currentSource }
-          }
-          const { code } = await jsHandler(currentSource, runtimeSet, handlerOptions)
-          debug('js handle: %s', file.path)
-          return {
-            result: code,
-          }
-        },
-      })
-      rememberGulpProcessCacheKey(gulpProcessCacheKeys, file.path)
-      pruneGulpProcessCache(cache, gulpProcessCacheKeys)
-    }, () => resolveGulpTransformTimingDetails('js'))
-
-  const transformWxml = (options: Partial<ITemplateHandlerOptions> = {}) =>
-    createVinylTransform('html', async (file) => {
-      if (!file.contents) {
-        return
-      }
-      const rawSource = file.contents.toString()
-      await refreshRuntimeSetForSource(file, rawSource, 'html')
-      await runtimeState.readyPromise
-      await processCachedTask<string>({
-        cache,
-        cacheKey: file.path,
-        rawSource,
-        applyResult(source) {
-          file.contents = Buffer.from(source)
-        },
-        onCacheHit() {
-          debug('html cache hit: %s', file.path)
-        },
-        async transform() {
-          await runtimeState.readyPromise
-          const code = await templateHandler(rawSource, resolveWxmlHandlerOptions(options))
-          debug('html handle: %s', file.path)
-          return {
-            result: code,
-          }
-        },
-      })
-      rememberGulpProcessCacheKey(gulpProcessCacheKeys, file.path)
-      pruneGulpProcessCache(cache, gulpProcessCacheKeys)
-    }, () => resolveGulpTransformTimingDetails('html'))
-
-  return {
-    adaptWxss,
-    generateWxss,
-    transformWxss,
-    transformWxml,
-    transformJs,
-  }
+  return createGulpFileTransforms({
+    cache,
+    createRuntimeSetHash,
+    debug,
+    generatedCssPreflightModeByFile,
+    getSourceCandidateGetter: () => cachedGulpSourceCandidateGetter,
+    getSourceCandidateSourceGetter: () => cachedGulpSourceCandidateSourceGetter,
+    getRuntimeSet: () => runtimeSet,
+    gulpProcessCacheKeys,
+    opts,
+    refreshGulpV4SourceCandidates,
+    refreshRuntimeSet,
+    refreshRuntimeSetForSource,
+    registerAutoCssSource,
+    resolveGulpStyleOutputExtension,
+    resolveGulpTransformTimingDetails,
+    resolveModuleGraphOptions,
+    resolveWxmlHandlerOptions,
+    resolveWxssFileHandlerOptions,
+    resolveWxssUserHandlerOptions,
+    runtimeState,
+  })
 }
