+++ packages/weapp-tailwindcss/src/bundlers/webpack/BaseUnifiedPlugin/v5-assets.ts
@@ -3,72 +3,39 @@ import type { WebpackCssHandlerOptions } from './v5-assets/pipeline-helpers'
 import path from 'node:path'
 import process from 'node:process'
 import { MappingChars2String } from '@weapp-core/escape'
-import { filterExistingCssRules } from '@weapp-tailwindcss/postcss'
 import { pluginName } from '@/constants'
 import { normalizeWeappTailwindcssGeneratorOptions } from '@/generator'
-import { shouldSkipJsTransform } from '@/js/precheck'
 import { ensureRuntimeClassSet } from '@/tailwindcss/runtime'
 import { getRuntimeClassSetSignature } from '@/tailwindcss/runtime/cache'
 import { getGroupedEntries } from '@/utils'
-import { processCachedTask } from '../../shared/cache'
 import { annotateCssSourceTrace, createCssSourceTraceCacheSignature, isCssSourceTraceEnabled } from '../../shared/css-source-trace'
-import { hasBundlerGeneratedCssMarker, stripBundlerGeneratedCssMarkers } from '../../shared/generated-css-marker'
-import { hasTailwindGeneratedCss, hasTailwindGeneratedCssMarkers, hasTailwindSourceDirectives, isPureLocalCssImportWrapper } from '../../shared/generator-css'
-import { hasTailwindApplyDirective, hasTailwindRootDirectives, removeTailwindSourceDirectives } from '../../shared/generator-css/directives'
-import { removeGeneratedSelectorCompatCss } from '../../shared/generator-css/legacy-selectors'
-import { isCommentOnlyCss } from '../../shared/generator-css/user-css'
-import { emitHmrTiming } from '../../shared/hmr-timing'
-import { toAbsoluteOutputPath } from '../../shared/module-graph'
-import { pushConcurrentTaskFactories, resolveTaskConcurrency } from '../../shared/run-tasks'
-import { generateTailwindV4Css } from '../../shared/v4-generation-core'
+import { resolveTaskConcurrency } from '../../shared/run-tasks'
 import { createBundleBuildState, updateBundleBuildState } from '../../vite/bundle-state'
-import { createScopedGeneratorRuntime } from '../../vite/generate-bundle/scoped-generator'
 import { createCandidateSignature } from '../../vite/generate-bundle/signatures'
 import { createBundleRuntimeClassSetManager } from '../../vite/incremental-runtime-class-set'
 import { collectStrictEscapedRuntimeCandidates, createEscapeFragments } from '../../vite/incremental-runtime-class-set/escaped-candidates'
-import { resolveTailwindV4EntriesFromCssCached } from '../../vite/source-scan'
-import { isWebpackCssLoaderRuntimeSource } from '../shared/css-loader-runtime'
-import { createAssetHashByChunkMap, createRuntimeAwareCssHash, createWebpackCssAssetResourceMap, createWebpackDirectCssAssetResourceMap, getCacheKey } from './shared'
+import { createAssetHashByChunkMap, createWebpackCssAssetResourceMap, createWebpackDirectCssAssetResourceMap } from './shared'
 import { createWebpackCssSourceResolvers } from './v5-assets/css-source-resolvers'
+import { finalizeWebpackProcessAssets } from './v5-assets/finalize-assets'
+import { processWebpackGeneratedCssAsset } from './v5-assets/generated-css-task'
 import { buildWebpackBundleSnapshot, createWebpackAssetUpdater, releaseWebpackBundleSnapshotSources } from './v5-assets/helpers'
+import { enqueueWebpackHtmlAndJsTasks } from './v5-assets/html-js-tasks'
 import { applyWebpackLinkedJsResults, createWebpackJsAssetModuleGraph } from './v5-assets/js-module-graph'
 import {
   addRuntimeTransformCandidates,
   collectGeneratedCssRuntimeCandidates,
-  collectWebpackBareSelectorUserCss,
   collectWebpackJsRuntimeCandidatesFromAssets,
   collectWebpackJsRuntimeTokenSignature,
   createWebpackCssSourceTraceTokenSources,
-  createWebpackCurrentAssetUserRawSource,
-  createWebpackGeneratorUserCssSourceAppend,
-  createWebpackUserCssSourceAppend,
   finalizeTracedWebpackCssAsset,
   finalizeWebpackCssAssetOutputSource,
   finalizeWebpackCssAssetSource,
   getRuntimeClassSetSync,
-  hasAdditionalWebpackAssetUserCssMarkers,
-  hasDeferredWebpackGeneratedCss,
-  hasMissingRuntimeCandidates,
-  hasProcessedCssAssetUrl,
-  hasUsableWebpackGeneratorCssSources,
   isRuntimeTransformCandidate,
-  normalizeWebpackGeneratorCssSources,
-  pruneWebpackCssHandlerOptionCaches,
-  removeWebpackGeneratorNonTailwindImports,
-  removeWebpackTailwindGeneratedAssetCss,
-  resolveGeneratedCssRuntimeCandidates,
   resolveWebpackCssAssetModuleResource,
-  resolveWebpackGeneratorRawSource,
-  resolveWebpackMemoryDebugStats,
-  scopeWebpackGeneratorOptionsToCssSource,
-  shouldAppendCurrentWebpackAssetUserCss,
-  shouldConsumeWebpackLoaderGeneratedCss,
-  shouldFallbackToWebpackUserCssOnGeneratorError,
-  shouldUseWebpackAssetAsGeneratorUserCss,
-  stringifyOptionalWebpackSourceValue,
   stringifyWebpackSourceLike,
-  stripTrailingLineWhitespace,
 } from './v5-assets/pipeline-helpers'
+import { processWebpackProcessedCssAsset } from './v5-assets/processed-css-task'
 import { createWebpackSourceCandidateScanCache } from './v5-assets/source-candidate-cache'
 import { refreshWebpackSourceCandidates } from './v5-assets/source-candidate-refresh'
 
@@ -407,7 +374,6 @@ export function setupWebpackV5ProcessAssetsHook(options: SetupWebpackV5ProcessAs
         }
         debug('get runtimeSet, class count: %d, transform class count: %d', runtimeSet.size, transformRuntimeSet.size)
         const tasks: Promise<void>[] = []
-        const htmlTaskFactories: Array<() => Promise<void>> = []
         const cssTaskFactories: Array<() => Promise<void>> = []
         const enqueueTask = async (
           factory: () => Promise<void>,
@@ -429,932 +395,57 @@ export function setupWebpackV5ProcessAssetsHook(options: SetupWebpackV5ProcessAs
           }
           target.push(timedFactory)
         }
-        if (!isWebGeneratorTarget && Array.isArray(groupedEntries.html)) {
-          for (const element of groupedEntries.html) {
-            const [file, originalSource] = element
-            let rawSource: string | undefined
-            const readRawSource = () => {
-              rawSource ??= originalSource.source().toString()
-              return rawSource
-            }
-
-            const cacheKey = file
-            const hashKey = `${file}:asset`
-            rememberProcessCacheKey(cacheKey, hashKey)
-            const chunkHash = assetHashByChunk.get(file)
-            await enqueueTask(async () => {
-              await processCachedTask({
-                cache: compilerOptions.cache,
-                cacheKey,
-                hashKey,
-                rawSource: chunkHash === undefined ? readRawSource() : undefined,
-                hash: chunkHash,
-                applyResult(source, { cacheHit }) {
-                  updateAssetIfChanged(file, source, {
-                    compare: !cacheHit,
-                    notifyUpdate: !cacheHit,
-                  })
-                },
-                onCacheHit() {
-                  debug('html cache hit: %s', file)
-                },
-                transform: async () => {
-                  const wxml = await compilerOptions.templateHandler(readRawSource(), defaultTemplateHandlerOptions)
-                  const source = new ConcatSource(wxml)
-                  debug('html handle: %s', file)
-
-                  return {
-                    result: source,
-                  }
-                },
-              })
-            }, htmlTaskFactories, 'tasks.html')
-          }
-        }
-
-        const jsTaskFactories: Array<() => Promise<void>> = []
-        const enqueueJsTask = async (factory: () => Promise<void>) => {
-          await enqueueTask(factory, jsTaskFactories, 'tasks.js')
-        }
-
-        if (!isWebGeneratorTarget && Array.isArray(groupedEntries.js)) {
-          for (const [file] of groupedEntries.js) {
-            const cacheKey = getCacheKey(file)
-            const asset = compilation.getAsset(file)
-            if (!asset) {
-              continue
-            }
-            const hashKey = `${file}:asset`
-            rememberProcessCacheKey(cacheKey, hashKey)
-            const absoluteFile = toAbsoluteOutputPath(file, outputDir)
-            const initialSource = asset.source.source()
-            const initialRawSource = typeof initialSource === 'string' ? initialSource : initialSource.toString()
-            const chunkHash = assetHashByChunk.get(file)
-            await enqueueJsTask(async () => {
-              await processCachedTask({
-                cache: compilerOptions.cache,
-                cacheKey,
-                hashKey,
-                rawSource: chunkHash === undefined ? initialRawSource : undefined,
-                hash: chunkHash,
-                applyResult(source, { cacheHit }) {
-                  const updated = updateAssetIfChanged(file, source, {
-                    compare: !cacheHit,
-                    notifyUpdate: !cacheHit,
-                  })
-                  if (updated) {
-                    rememberTransformedRuntimeCandidates(source)
-                  }
-                },
-                onCacheHit() {
-                  debug('js cache hit: %s', file)
-                },
-                transform: async () => {
-                  const currentAsset = compilation.getAsset(file)
-                  const currentSourceValue = currentAsset?.source.source()
-                  const currentSource = stringifyOptionalWebpackSourceValue(currentSourceValue)
-                  const handlerOptions = {
-                    tailwindcssMajorVersion: runtimeState.tailwindRuntime.majorVersion,
-                    generateMap: false,
-                    filename: absoluteFile,
-                    moduleGraph: moduleGraphOptions,
-                    babelParserOptions: {
-                      sourceFilename: absoluteFile,
-                    },
-                  }
-                  if (shouldSkipJsTransform(currentSource, {
-                    ...handlerOptions,
-                    classNameSet: transformRuntimeSet,
-                  })) {
-                    return { result: new ConcatSource(currentSource) }
-                  }
-                  const { code, linked } = await compilerOptions.jsHandler(currentSource, transformRuntimeSet, handlerOptions)
-                  const source = new ConcatSource(code)
-                  debug('js handle: %s', file)
-                  applyWebpackLinkedJsResults({
-                    ConcatSource,
-                    compilation,
-                    compilerOptions,
-                    debug,
-                    jsAssets,
-                    linked,
-                  })
-                  return {
-                    result: source,
-                  }
-                },
-              })
-            })
-          }
-        }
-
+        const { htmlTaskFactories, jsTaskFactories } = await enqueueWebpackHtmlAndJsTasks({
+          ConcatSource,
+          applyWebpackLinkedJsResults,
+          assetHashByChunk,
+          compilation,
+          compilerOptions,
+          debug,
+          defaultTemplateHandlerOptions,
+          enqueueTask,
+          groupedEntries,
+          isWebGeneratorTarget,
+          jsAssets,
+          moduleGraphOptions,
+          outputDir,
+          rememberProcessCacheKey,
+          rememberTransformedRuntimeCandidates,
+          runtimeState,
+          transformRuntimeSet,
+          updateAssetIfChanged,
+        })
+        const cssTaskContext = { ConcatSource, assetHashByChunk, compilation, compilerOptions, configuredCssEntryFiles, configuredMainCssEntryFiles, createRuntimeSetHash, cssSourceTraceSignature, cssSources, cssTaskFactories, debug, enqueueTask, finalizeCssAssetSource, finalizeTracedCss, generatedCssSources, generatorRuntimeSet, getCssHandlerOptions, getCssUserHandlerOptions, getGeneratorRuntimeSet, hasConfiguredTailwindV4SourceRoots, isKnownWebpackProcessedCssAsset, isSameWebpackSourceScope, isWebGeneratorTarget, isWebpackProcessedCssAsset, processedCssAssetSkipDecisionCache, rememberProcessCacheKey, resolveWebpackCssSourceFile, runtimeAffectingSourceHash, runtimeState, transformRuntimeSet, updateAssetIfChanged, watchMode, webpackSourceCandidateSet, webpackSourceCandidateValueSignature, webpackSourceCandidates }
         for (const element of groupedEntries.css) {
-          const [file, originalSource] = element
-
-          let rawSource: string | undefined
-          const readRawSource = () => {
-            rawSource ??= originalSource.source().toString()
-            return rawSource
-          }
-          const chunkHash = assetHashByChunk.get(file)
-          const cssHandlerOptionsForProcessedAsset = getCssHandlerOptions(file, readRawSource())
-          const processedCssAssetMetadata = {
-            isMainCssChunk: cssHandlerOptionsForProcessedAsset.isMainChunk,
-          }
-          const processedSourceFile = cssHandlerOptionsForProcessedAsset.sourceOptions?.sourceFile
-          const processedSourceCss = processedSourceFile ? cssSources.get(path.resolve(processedSourceFile))?.css : undefined
-          const shouldRegenerateProcessedTailwindV4SourceCss = processedSourceCss !== undefined
-            && (
-              hasTailwindSourceDirectives(processedSourceCss, { importFallback: true })
-              || processedSourceCss.includes('@config')
-            )
-          const processedCssAssetKnown = isKnownWebpackProcessedCssAsset?.(file, processedCssAssetMetadata) === true
-          const processedLoaderGeneratedCss = processedSourceFile
-            ? generatedCssSources.get(path.resolve(processedSourceFile))
-            : undefined
-          const processedAssetSourceHash = watchMode
-            && isWebGeneratorTarget
-            && cssHandlerOptionsForProcessedAsset.isMainChunk
-            ? compilerOptions.cache.computeHash(readRawSource())
-            : chunkHash === undefined
-              ? processedCssAssetKnown
-                ? 'webpack-css-asset:known'
-                : compilerOptions.cache.computeHash(readRawSource())
-              : 'webpack-css-asset:chunk'
-          const processedCssHashKey = createRuntimeAwareCssHash(
-            chunkHash,
-            processedAssetSourceHash,
-            `${createRuntimeSetHash(getGeneratorRuntimeSet())}:${runtimeAffectingSourceHash}:${webpackSourceCandidates?.signatureHash ?? 'source-candidates:0'}:${webpackSourceCandidateValueSignature}:${cssSourceTraceSignature}`,
-          )
-          const processedCssDecisionCacheKey = `${file}:${processedCssHashKey}`
-          let currentProcessedRawSource: string | undefined
-          let hasGeneratedCssMarker = false
-          let hasTailwindGeneratedAssetCss = false
-          const readCurrentProcessedRawSource = () => {
-            currentProcessedRawSource ??= readRawSource()
-            return currentProcessedRawSource
-          }
-          const cachedSkipProcessedCssAsset = processedCssAssetKnown
-            ? processedCssAssetSkipDecisionCache.get(processedCssDecisionCacheKey)
-            : undefined
-          const shouldRegenerateStaleProcessedWebCssAsset = isWebGeneratorTarget
-            && !processedCssAssetKnown
-            && cachedSkipProcessedCssAsset === undefined
-            && cssHandlerOptionsForProcessedAsset.isMainChunk
-            && webpackSourceCandidateSet !== undefined
-            && (
-              hasMissingRuntimeCandidates(processedLoaderGeneratedCss?.classSet, webpackSourceCandidateSet)
-              || hasMissingRuntimeCandidates(
-                resolveGeneratedCssRuntimeCandidates(
-                  readCurrentProcessedRawSource(),
-                  processedLoaderGeneratedCss?.classSet,
-                ),
-                webpackSourceCandidateSet,
-              )
-            )
-          if (cachedSkipProcessedCssAsset !== undefined) {
-            hasGeneratedCssMarker = cachedSkipProcessedCssAsset && cssHandlerOptionsForProcessedAsset.isMainChunk
-            hasTailwindGeneratedAssetCss = hasGeneratedCssMarker
-          }
-          else {
-            const source = readCurrentProcessedRawSource()
-            hasGeneratedCssMarker = hasBundlerGeneratedCssMarker(source)
-            hasTailwindGeneratedAssetCss = hasTailwindGeneratedCss(source)
-              || hasTailwindGeneratedCssMarkers(source)
-          }
-          const hasProcessedAssetTailwindDirectives = () => {
-            const source = readCurrentProcessedRawSource()
-            return hasTailwindRootDirectives(source, { importFallback: true })
-              || hasTailwindSourceDirectives(source, { importFallback: true })
-              || hasTailwindApplyDirective(source)
-          }
-          const shouldForceConfiguredMainCssGeneration = cssHandlerOptionsForProcessedAsset.isMainChunk
-            && hasConfiguredTailwindV4SourceRoots()
-            && !hasGeneratedCssMarker
-            && (
-              configuredMainCssEntryFiles.length > 0
-              || shouldRegenerateProcessedTailwindV4SourceCss
-              || hasProcessedAssetTailwindDirectives()
-            )
-          const hasProcessedMainAssetUserCss = cachedSkipProcessedCssAsset === undefined
-            && cssHandlerOptionsForProcessedAsset.isMainChunk
-            && (hasGeneratedCssMarker || hasTailwindGeneratedAssetCss)
-            && createWebpackUserCssSourceAppend(
-              [...cssSources.entries()].map(([sourceFile, source]) => ({
-                ...source,
-                file: sourceFile,
-              })),
-              readCurrentProcessedRawSource(),
-            ) !== undefined
-          const hasProcessedLoaderGeneratedUserCss = cachedSkipProcessedCssAsset === undefined
-            && processedLoaderGeneratedCss !== undefined
-            && hasAdditionalWebpackAssetUserCssMarkers(
-              processedLoaderGeneratedCss.css,
-              readCurrentProcessedRawSource(),
-            )
-          const shouldFinalizeProcessedWebCssAsset = isWebGeneratorTarget
-            && !shouldForceConfiguredMainCssGeneration
-            && !shouldRegenerateProcessedTailwindV4SourceCss
-            && hasTailwindSourceDirectives(readCurrentProcessedRawSource(), { importFallback: true })
-          const shouldPreserveFinalWebCssAsset = isWebGeneratorTarget
-            && processedSourceFile === undefined
-            && !shouldForceConfiguredMainCssGeneration
-            && (hasGeneratedCssMarker || hasTailwindGeneratedAssetCss)
-          const shouldSkipKnownProcessedCssAsset = !shouldForceConfiguredMainCssGeneration
-            && !shouldRegenerateProcessedTailwindV4SourceCss
-            && !shouldRegenerateStaleProcessedWebCssAsset
-            && (
-              processedCssAssetKnown
-              || isWebpackProcessedCssAsset?.(file, readCurrentProcessedRawSource(), processedCssAssetMetadata)
-            )
-            && !hasProcessedMainAssetUserCss
-            && !hasProcessedLoaderGeneratedUserCss
-            && (!cssHandlerOptionsForProcessedAsset.isMainChunk || hasGeneratedCssMarker || hasTailwindGeneratedAssetCss)
-          const shouldSkipProcessedCssAsset = (
-            cachedSkipProcessedCssAsset
-            ?? (shouldFinalizeProcessedWebCssAsset || shouldPreserveFinalWebCssAsset || shouldSkipKnownProcessedCssAsset)
-          )
-          if (processedCssAssetKnown && cachedSkipProcessedCssAsset === undefined && !shouldFinalizeProcessedWebCssAsset && !shouldPreserveFinalWebCssAsset) {
-            processedCssAssetSkipDecisionCache.set(processedCssDecisionCacheKey, shouldSkipProcessedCssAsset === true)
-          }
-          if (shouldSkipProcessedCssAsset) {
-            const hashKey = `${file}:asset`
-            const sourceHash = processedAssetSourceHash
-            rememberProcessCacheKey(file, hashKey)
-            await enqueueTask(async () => {
-              await processCachedTask({
-                cache: compilerOptions.cache,
-                cacheKey: file,
-                hashKey,
-                rawSource: chunkHash === undefined && !processedCssAssetKnown
-                  ? readCurrentProcessedRawSource()
-                  : undefined,
-                hash: createRuntimeAwareCssHash(
-                  chunkHash,
-                  sourceHash,
-                  `${createRuntimeSetHash(getGeneratorRuntimeSet())}:${runtimeAffectingSourceHash}:${webpackSourceCandidates?.signatureHash ?? 'source-candidates:0'}:${webpackSourceCandidateValueSignature}:${cssSourceTraceSignature}`,
-                ),
-                applyResult(source, { cacheHit }) {
-                  updateAssetIfChanged(file, source, {
-                    compare: !cacheHit,
-                    notifyUpdate: !cacheHit,
-                  })
-                },
-                onCacheHit() {
-                  debug('css webpack-loader-pipeline cache hit: %s', file)
-                },
-                transform: async () => {
-                  const source = readCurrentProcessedRawSource()
-                  const missingProcessedLoaderGeneratedCss = isWebGeneratorTarget && processedLoaderGeneratedCss
-                    ? filterExistingCssRules(
-                        source,
-                        stripBundlerGeneratedCssMarkers(processedLoaderGeneratedCss.css),
-                      )
-                    : ''
-                  const sourceWithLoaderGeneratedCss = missingProcessedLoaderGeneratedCss.trim().length === 0
-                    ? source
-                    : createWebpackGeneratorUserCssSourceAppend(
-                      { css: source, processed: true },
-                      { css: missingProcessedLoaderGeneratedCss, processed: true },
-                    )!.css
-                  const processedBareSelectorSourceCss = processedSourceCss
-                    ?? (hasTailwindGeneratedAssetCss ? removeWebpackTailwindGeneratedAssetCss(sourceWithLoaderGeneratedCss) : undefined)
-                  const shouldTransformGeneratedAssetCss = hasTailwindGeneratedAssetCss
-                    && (
-                      !hasGeneratedCssMarker
-                      || (
-                        !isWebGeneratorTarget
-                        && hasDeferredWebpackGeneratedCss(
-                          source,
-                          [...generatedCssSources.values()].map(item => item.classSet),
-                        )
-                      )
-                    )
-                  const handledCss = shouldTransformGeneratedAssetCss
-                    ? isWebGeneratorTarget
-                      ? sourceWithLoaderGeneratedCss
-                      : (await compilerOptions.styleHandler(
-                          sourceWithLoaderGeneratedCss,
-                          cssHandlerOptionsForProcessedAsset,
-                        )).css
-                    : sourceWithLoaderGeneratedCss
-                  const nextCss = stripTrailingLineWhitespace(finalizeCssAssetSource(handledCss, {
-                    cssPreflight: cssHandlerOptionsForProcessedAsset.isMainChunk,
-                    generatedCss: hasGeneratedCssMarker || hasTailwindGeneratedAssetCss,
-                  }))
-                  const processedSourceBareUserCss = isWebGeneratorTarget || processedBareSelectorSourceCss === undefined
-                    ? undefined
-                    : createWebpackGeneratorUserCssSourceAppend({
-                        css: collectWebpackBareSelectorUserCss(processedBareSelectorSourceCss),
-                        processed: false,
-                      })
-                  const finalizedProcessedSourceBareUserCss = processedSourceBareUserCss === undefined
-                    ? ''
-                    : finalizeCssAssetSource(processedSourceBareUserCss.css, {
-                        cssPreflight: false,
-                        generatedCss: false,
-                      })
-                  const missingProcessedSourceBareUserCss = finalizedProcessedSourceBareUserCss.trim().length === 0
-                    ? ''
-                    : filterExistingCssRules(nextCss, finalizedProcessedSourceBareUserCss)
-                  const css = missingProcessedSourceBareUserCss.trim().length === 0
-                    ? nextCss
-                    : createWebpackGeneratorUserCssSourceAppend(
-                      {
-                        css: nextCss,
-                        processed: true,
-                      },
-                      {
-                        css: missingProcessedSourceBareUserCss,
-                        processed: true,
-                      },
-                    )!.css
-                  debug('css skip webpack-loader-pipeline asset: %s', file)
-                  return {
-                    result: new ConcatSource(finalizeTracedCss(css, cssHandlerOptionsForProcessedAsset, { finalized: true })),
-                  }
-                },
-              })
-            }, cssTaskFactories, 'tasks.css')
+          if (await processWebpackProcessedCssAsset(element, cssTaskContext)) {
             continue
           }
-          const currentRawSource = readRawSource()
-          if (isWebpackCssLoaderRuntimeSource(currentRawSource)) {
-            const hashKey = `${file}:asset`
-            rememberProcessCacheKey(file, hashKey)
-            await enqueueTask(async () => {
-              await processCachedTask({
-                cache: compilerOptions.cache,
-                cacheKey: file,
-                hashKey,
-                rawSource: currentRawSource,
-                hash: createRuntimeAwareCssHash(
-                  chunkHash,
-                  compilerOptions.cache.computeHash(currentRawSource),
-                  'webpack-css-loader-runtime',
-                ),
-                applyResult(source, { cacheHit }) {
-                  updateAssetIfChanged(file, source, {
-                    compare: !cacheHit,
-                    notifyUpdate: !cacheHit,
-                  })
-                },
-                onCacheHit() {
-                  debug('css-loader runtime cache hit: %s', file)
-                },
-                transform: async () => ({
-                  result: new ConcatSource(currentRawSource),
-                }),
-              })
-            }, cssTaskFactories, 'tasks.css')
-            continue
-          }
-          const cacheKey = file
-          const hashKey = `${file}:asset`
-          rememberProcessCacheKey(cacheKey, hashKey)
-          const cssHandlerOptionsForHash = getCssHandlerOptions(file, currentRawSource)
-          const cssChunkHash = watchMode
-            && cssHandlerOptionsForHash.isMainChunk
-            ? undefined
-            : chunkHash
-          const cssSourceHash = (() => {
-            const sourceFile = resolveWebpackCssSourceFile(file, currentRawSource)
-            const sourceCss = sourceFile ? cssSources.get(sourceFile)?.css : undefined
-            const generatorSourceCss = removeWebpackGeneratorNonTailwindImports(sourceCss)
-            if (sourceCss === undefined) {
-              return sourceFile === undefined
-                ? 'webpack-css-source:0'
-                : `webpack-css-source:0:${sourceFile}`
-            }
-            return `webpack-css-source:1:${compilerOptions.cache.computeHash(sourceCss)}:${generatorSourceCss === sourceCss || generatorSourceCss === undefined ? 'generator-source:0' : compilerOptions.cache.computeHash(generatorSourceCss)}`
-          })()
-          const runtimeAwareHash = createRuntimeAwareCssHash(
-            cssChunkHash,
-            compilerOptions.cache.computeHash(currentRawSource),
-            `${createRuntimeSetHash(getGeneratorRuntimeSet())}:${runtimeAffectingSourceHash}:${webpackSourceCandidates?.signatureHash ?? 'source-candidates:0'}:${webpackSourceCandidateValueSignature}:${cssSourceTraceSignature}:${cssSourceHash}`,
-          )
-          await enqueueTask(async () => {
-            await processCachedTask({
-              cache: compilerOptions.cache,
-              cacheKey,
-              hashKey,
-              rawSource: currentRawSource,
-              hash: runtimeAwareHash,
-              applyResult(source, { cacheHit }) {
-                updateAssetIfChanged(file, source, {
-                  compare: !cacheHit,
-                  notifyUpdate: !cacheHit,
-                })
-              },
-              onCacheHit() {
-                debug('css cache hit: %s', file)
-              },
-              transform: async () => {
-                await runtimeState.readyPromise
-                const cssHandlerOptions = getCssHandlerOptions(file, currentRawSource)
-                const generatorRawSource = resolveWebpackGeneratorRawSource(currentRawSource, cssHandlerOptions)
-                if (isWebpackCssLoaderRuntimeSource(generatorRawSource)) {
-                  return {
-                    result: new ConcatSource(currentRawSource),
-                  }
-                }
-                const sourceFile = cssHandlerOptions.sourceOptions?.sourceFile
-                const sourceCss = sourceFile ? cssSources.get(path.resolve(sourceFile))?.css : undefined
-                const isConfiguredCssSource = sourceFile !== undefined
-                  && configuredCssEntryFiles.some(entry => path.resolve(entry) === path.resolve(sourceFile))
-                const isConfiguredMainCssSource = sourceFile !== undefined
-                  && configuredMainCssEntryFiles.some(entry => path.resolve(entry) === path.resolve(sourceFile))
-                const sourceCssHasTailwindRoot = sourceCss !== undefined
-                  && hasTailwindRootDirectives(sourceCss, { importFallback: true })
-                const shouldPreserveExistingPreflight = cssHandlerOptions.isMainChunk
-                  || (!isConfiguredMainCssSource && (isConfiguredCssSource || sourceCssHasTailwindRoot))
-                const loaderGeneratedCss = sourceFile
-                  ? generatedCssSources.get(path.resolve(sourceFile))
-                  : undefined
-                const shouldRegenerateExplicitTailwindV4CssSource = sourceCss !== undefined
-                  && (
-                    hasTailwindSourceDirectives(sourceCss, { importFallback: true })
-                    || sourceCss.includes('@config')
-                  )
-                const explicitTailwindV4SourceCandidates = shouldRegenerateExplicitTailwindV4CssSource
-                  && sourceCss
-                  && sourceFile
-                  && webpackSourceCandidates?.getSourceCandidatesForEntries
-                  ? await resolveTailwindV4EntriesFromCssCached(sourceCss, path.dirname(sourceFile))
-                      .then(resolved => resolved?.entries
-                        ? webpackSourceCandidates.getSourceCandidatesForEntries(resolved.entries)
-                        : undefined)
-                  : undefined
-                if (
-                  loaderGeneratedCss
-                  && shouldConsumeWebpackLoaderGeneratedCss({
-                    allowMarkerlessRegistryMatch: configuredCssEntryFiles.length === 1,
-                    hasBundlerGeneratedCssMarker: hasBundlerGeneratedCssMarker(currentRawSource),
-                    loaderGeneratedClassSet: loaderGeneratedCss.classSet,
-                    sourceCandidates: explicitTailwindV4SourceCandidates,
-                    shouldRegenerateExplicitTailwindV4CssSource,
-                    watchMode,
-                  })
-                ) {
-                  for (const className of loaderGeneratedCss.classSet) {
-                    generatorRuntimeSet.add(className)
-                    transformRuntimeSet.add(className)
-                  }
-                  for (const dependency of loaderGeneratedCss.dependencies) {
-                    compilation.fileDependencies?.add?.(dependency)
-                  }
-                  const currentRawSourceWithoutBundlerMarkers = stripBundlerGeneratedCssMarkers(currentRawSource)
-                  const currentAssetHasProcessedUrl = hasProcessedCssAssetUrl(currentRawSourceWithoutBundlerMarkers)
-                    && currentRawSourceWithoutBundlerMarkers !== loaderGeneratedCss.css
-                  const currentAssetUserCss = currentAssetHasProcessedUrl
-                    ? currentRawSourceWithoutBundlerMarkers
-                    : shouldUseWebpackAssetAsGeneratorUserCss(currentRawSourceWithoutBundlerMarkers, loaderGeneratedCss.css, {
-                      processed: true,
-                    })
-                      ? removeGeneratedSelectorCompatCss(
-                          currentRawSourceWithoutBundlerMarkers,
-                          loaderGeneratedCss.css,
-                        )
-                      : undefined
-                  const loaderGeneratedCssWithoutMarkers = stripBundlerGeneratedCssMarkers(loaderGeneratedCss.css)
-                  const currentAssetUserCssWithoutMarkers = currentAssetUserCss === undefined
-                    ? ''
-                    : stripBundlerGeneratedCssMarkers(currentAssetUserCss)
-                  const currentAssetUserCssHasRules = currentAssetUserCssWithoutMarkers.trim().length > 0
-                    && !isCommentOnlyCss(currentAssetUserCssWithoutMarkers)
-                  const currentAssetMissingUserCss = !currentAssetUserCssHasRules
-                    ? ''
-                    : filterExistingCssRules(currentAssetUserCssWithoutMarkers, loaderGeneratedCssWithoutMarkers)
-                  if (
-                    isConfiguredMainCssSource
-                    && !cssHandlerOptions.isMainChunk
-                  ) {
-                    debug('css skip duplicate webpack loader main generation: %s <- %s', file, sourceFile)
-                    const userCss = currentAssetMissingUserCss.trim().length === 0 || isCommentOnlyCss(currentAssetMissingUserCss)
-                      ? ''
-                      : finalizeCssAssetSource(currentAssetMissingUserCss, {
-                          cssPreflight: false,
-                          generatedCss: false,
-                        })
-                    return {
-                      result: new ConcatSource(finalizeTracedCss(userCss, cssHandlerOptions)),
-                    }
-                  }
-                  const mergedLoaderCss = currentAssetUserCss === undefined
-                    ? loaderGeneratedCss.css
-                    : (createWebpackGeneratorUserCssSourceAppend(
-                        {
-                          css: currentAssetUserCss === undefined
-                            ? loaderGeneratedCss.css
-                            : currentAssetHasProcessedUrl
-                              ? removeGeneratedSelectorCompatCss(loaderGeneratedCss.css, currentAssetUserCss)
-                              : filterExistingCssRules(currentAssetUserCss, loaderGeneratedCss.css),
-                          processed: true,
-                        },
-                        currentAssetUserCss === undefined
-                          ? undefined
-                          : {
-                              css: currentAssetUserCss,
-                              processed: true,
-                            },
-                      )!.css)
-                  const handledLoaderCss = isWebGeneratorTarget
-                    ? mergedLoaderCss
-                    : (await compilerOptions.styleHandler(mergedLoaderCss, cssHandlerOptions)).css
-                  const finalizedLoaderCss = finalizeCssAssetSource(handledLoaderCss, {
-                    cssPreflight: cssHandlerOptions.isMainChunk,
-                    generatedCss: true,
-                    preserveExistingPreflight: shouldPreserveExistingPreflight,
-                  })
-                  const loaderSourceBareUserCss = isWebGeneratorTarget
-                    ? undefined
-                    : createWebpackGeneratorUserCssSourceAppend(
-                        ...[
-                          currentAssetUserCss === undefined
-                            ? undefined
-                            : {
-                                css: currentAssetUserCss,
-                                processed: true,
-                              },
-                          sourceCss === undefined
-                            ? undefined
-                            : {
-                                css: sourceCss,
-                                processed: false,
-                              },
-                        ].map((source) => {
-                          if (source === undefined) {
-                            return undefined
-                          }
-                          return {
-                            css: collectWebpackBareSelectorUserCss(source.css),
-                            processed: source.processed,
-                          }
-                        }),
-                      )
-                  const handledLoaderSourceBareUserCss = loaderSourceBareUserCss === undefined
-                    ? ''
-                    : loaderSourceBareUserCss.processed
-                      ? loaderSourceBareUserCss.css
-                      : (await compilerOptions.styleHandler(loaderSourceBareUserCss.css, cssHandlerOptions)).css
-                  const finalizedLoaderSourceBareUserCss = handledLoaderSourceBareUserCss.trim().length === 0
-                    ? ''
-                    : finalizeCssAssetSource(handledLoaderSourceBareUserCss, {
-                        cssPreflight: false,
-                        generatedCss: false,
-                      })
-                  const missingLoaderSourceBareUserCss = finalizedLoaderSourceBareUserCss.trim().length === 0
-                    ? ''
-                    : filterExistingCssRules(finalizedLoaderCss, finalizedLoaderSourceBareUserCss)
-                  const css = finalizeTracedCss(
-                    missingLoaderSourceBareUserCss.trim().length === 0
-                      ? finalizedLoaderCss
-                      : createWebpackGeneratorUserCssSourceAppend(
-                        {
-                          css: finalizedLoaderCss,
-                          processed: true,
-                        },
-                        {
-                          css: missingLoaderSourceBareUserCss,
-                          processed: true,
-                        },
-                      )!.css,
-                    cssHandlerOptions,
-                    { finalized: true },
-                  )
-                  debug('css consume webpack loader generation: %s <- %s', file, sourceFile)
-                  return {
-                    result: new ConcatSource(css),
-                  }
-                }
-                const sourceCssProcessed = sourceFile ? cssSources.get(path.resolve(sourceFile))?.processed === true : false
-                const registeredUserRawSource = createWebpackUserCssSourceAppend(
-                  [...cssSources.entries()].map(([registeredSourceFile, source]) => ({
-                    ...source,
-                    file: registeredSourceFile,
-                  })),
-                  generatorRawSource,
-                  sourceFile,
-                  registeredSourceFile => isSameWebpackSourceScope(file, registeredSourceFile, sourceFile),
-                )
-                const currentAssetLooksGenerated = hasTailwindGeneratedCss(currentRawSource)
-                  || hasTailwindGeneratedCssMarkers(currentRawSource)
-                const currentAssetHasBundlerGeneratedMarker = hasBundlerGeneratedCssMarker(currentRawSource)
-                const currentAssetUserCssSource = currentAssetLooksGenerated
-                  ? removeWebpackTailwindGeneratedAssetCss(currentRawSource)
-                  : currentRawSource
-                if (
-                  sourceFile === undefined
-                  && currentAssetHasBundlerGeneratedMarker
-                  && (
-                    currentAssetUserCssSource.trim().length === 0
-                    || isCommentOnlyCss(currentAssetUserCssSource)
-                  )
-                ) {
-                  return {
-                    result: new ConcatSource(finalizeTracedCss('', cssHandlerOptions)),
-                  }
-                }
-                const currentAssetHasAdditionalUserCss = currentAssetLooksGenerated
-                  && (
-                    hasAdditionalWebpackAssetUserCssMarkers(currentAssetUserCssSource, generatorRawSource)
-                    || currentAssetUserCssSource.trim().length > 0
-                  )
-                const shouldPreserveGeneratedWebAssetUserCss = isWebGeneratorTarget
-                  && currentAssetLooksGenerated
-                  && !currentAssetHasBundlerGeneratedMarker
-                  && !currentAssetHasAdditionalUserCss
-                const hasExplicitSourceCssForCurrentAsset = sourceCss !== undefined
-                  && (
-                    hasTailwindRootDirectives(sourceCss, { importFallback: true })
-                    || hasTailwindSourceDirectives(sourceCss, { importFallback: true })
-                    || hasTailwindApplyDirective(sourceCss)
-                  )
-                const currentAssetHasUserCss = (sourceCssProcessed || hasExplicitSourceCssForCurrentAsset) && currentAssetLooksGenerated && !shouldPreserveGeneratedWebAssetUserCss
-                  ? currentAssetHasAdditionalUserCss
-                  : shouldUseWebpackAssetAsGeneratorUserCss(currentAssetUserCssSource, generatorRawSource, {
-                      processed: sourceCssProcessed || shouldPreserveGeneratedWebAssetUserCss,
-                    })
-                const shouldAppendCurrentAssetUserCss = shouldAppendCurrentWebpackAssetUserCss({
-                  currentAssetHasBundlerGeneratedMarker,
-                  currentAssetHasUserCss,
-                  currentAssetLooksGenerated,
-                  registeredUserRawSource,
-                  shouldPreserveGeneratedWebAssetUserCss,
-                  sourceCssProcessed,
-                })
-                const userRawSource = createWebpackGeneratorUserCssSourceAppend(
-                  createWebpackCurrentAssetUserRawSource({
-                    currentAssetHasUserCss,
-                    currentAssetLooksGenerated,
-                    currentAssetUserCssSource,
-                    shouldAppendCurrentAssetUserCss,
-                    sourceCssProcessed,
-                  }),
-                  registeredUserRawSource,
-                )
-                if (isPureLocalCssImportWrapper(currentRawSource)) {
-                  return {
-                    result: new ConcatSource(removeTailwindSourceDirectives(
-                      stripBundlerGeneratedCssMarkers(currentRawSource),
-                      { importFallback: true },
-                    )),
-                  }
-                }
-                const fallbackGeneratorRuntimeSet = getGeneratorRuntimeSet()
-                const hasExplicitTailwindV4SourceCss = sourceCss !== undefined
-                  && (
-                    hasTailwindSourceDirectives(sourceCss, { importFallback: true })
-                    || sourceCss.includes('@config')
-                  )
-                const currentAssetHasTailwindDirectives = hasTailwindRootDirectives(generatorRawSource, { importFallback: true })
-                  || hasTailwindSourceDirectives(generatorRawSource, { importFallback: true })
-                  || hasTailwindApplyDirective(generatorRawSource)
-                const shouldForceTailwindV4Generation = hasConfiguredTailwindV4SourceRoots()
-                  && (
-                    (configuredMainCssEntryFiles.length > 0 && sourceFile !== undefined)
-                    || currentAssetHasTailwindDirectives
-                    || hasExplicitTailwindV4SourceCss
-                  )
-                const shouldSkipUnmatchedMainCssGeneration = !isWebGeneratorTarget
-                  && cssHandlerOptions.isMainChunk
-                  && sourceFile === undefined
-                  && configuredMainCssEntryFiles.length > 0
-                  && !currentAssetHasTailwindDirectives
-                const shouldSkipPlainNonMainCssGeneration = !cssHandlerOptions.isMainChunk
-                  && sourceCss !== undefined
-                  && !hasExplicitSourceCssForCurrentAsset
-                  && !hasExplicitTailwindV4SourceCss
-                  && !currentAssetHasTailwindDirectives
-                const shouldMergeFallbackGeneratorRuntime = cssHandlerOptions.isMainChunk
-                  || currentAssetHasTailwindDirectives
-                  || hasExplicitSourceCssForCurrentAsset
-                const scopedFallbackGeneratorRuntimeSet = hasExplicitTailwindV4SourceCss || !shouldMergeFallbackGeneratorRuntime
-                  ? new Set<string>()
-                  : fallbackGeneratorRuntimeSet
-                const resolvedScopedGeneratorRuntimeSet = await createScopedGeneratorRuntime({
-                  cssHandlerOptions,
-                  fallbackRuntime: scopedFallbackGeneratorRuntimeSet,
-                  getSourceCandidatesForEntries: webpackSourceCandidates?.getSourceCandidatesForEntries,
-                  majorVersion: runtimeState.tailwindRuntime.majorVersion,
-                  outputFile: file,
-                  rawSource: sourceFile ? (cssSources.get(path.resolve(sourceFile))?.css ?? generatorRawSource) : generatorRawSource,
-                  shouldExcludeSubpackageSourceCandidates: () => false,
-                  sourceFile: sourceFile ?? file,
-                  scopedSourceCandidateGetter: webpackSourceCandidates?.getSourceCandidatesForEntries,
-                })
-                const scopedGeneratorRuntimeSet = !hasExplicitTailwindV4SourceCss
-                  && resolvedScopedGeneratorRuntimeSet !== scopedFallbackGeneratorRuntimeSet
-                  && shouldMergeFallbackGeneratorRuntime
-                  ? new Set([
-                      ...fallbackGeneratorRuntimeSet,
-                      ...resolvedScopedGeneratorRuntimeSet,
-                    ])
-                  : resolvedScopedGeneratorRuntimeSet
-                const generatorCssSources = normalizeWebpackGeneratorCssSources(cssHandlerOptions.sourceOptions?.cssSources)
-                const scopedGeneratorCompilerOptions = scopeWebpackGeneratorOptionsToCssSource(compilerOptions, sourceFile, {
-                  disableUnmatchedCssEntries: !isWebGeneratorTarget && cssHandlerOptions.isMainChunk,
-                })
-                const generatorCssHandlerOptions = generatorCssSources === undefined
-                  ? cssHandlerOptions
-                  : {
-                      ...cssHandlerOptions,
-                      sourceOptions: {
-                        ...cssHandlerOptions.sourceOptions,
-                        cssSources: generatorCssSources,
-                      },
-                    }
-                const generatorOptions = {
-                  opts: scopedGeneratorCompilerOptions,
-                  runtimeState,
-                  runtime: scopedGeneratorRuntimeSet,
-                  rawSource: generatorRawSource,
-                  forceGenerator: shouldForceTailwindV4Generation,
-                  ...(hasUsableWebpackGeneratorCssSources(generatorCssSources)
-                    ? { cssSources: generatorCssSources }
-                    : {}),
-                  ...(userRawSource === undefined ? {} : { userRawSource: userRawSource.css }),
-                  ...(userRawSource?.processed === true ? { userRawSourceProcessed: true } : {}),
-                  file,
-                  cssHandlerOptions: generatorCssHandlerOptions,
-                  cssUserHandlerOptions: getCssUserHandlerOptions(file),
-                  frameworkPostcssOwner: compilerOptions,
-                  frameworkPostcssStage: 'complete',
-                  getSourceCandidatesForEntries: webpackSourceCandidates?.getSourceCandidatesForEntries,
-                  sourceCandidates: scopedGeneratorRuntimeSet,
-                  restoreLocalCssImports: false,
-                  styleHandler: compilerOptions.styleHandler,
-                  debug,
-                }
-                let generated: Awaited<ReturnType<typeof generateTailwindV4Css>>
-                if (!shouldSkipUnmatchedMainCssGeneration && !shouldSkipPlainNonMainCssGeneration) {
-                  try {
-                    generated = await generateTailwindV4Css({
-                      ...generatorOptions,
-                      outputFile: file,
-                    })
-                  }
-                  catch (error) {
-                    const shouldFallbackToUserCss = shouldFallbackToWebpackUserCssOnGeneratorError({
-                      configuredMainCssEntryFilesLength: configuredMainCssEntryFiles.length,
-                      generatorRawSource,
-                      hasExplicitTailwindV4SourceCss,
-                    })
-                    if (!shouldFallbackToUserCss) {
-                      throw error
-                    }
-                    debug('css generator skipped for plain webpack css asset: %s %O', file, error)
-                    generated = undefined
-                  }
-                }
-                else {
-                  debug('css generator skipped for plain webpack css asset: %s', file)
-                }
-                const finalizedGeneratedCss = generated
-                  ? finalizeCssAssetSource(
-                      isWebGeneratorTarget && currentRawSource.includes('tailwindcss v4.')
-                        ? createWebpackGeneratorUserCssSourceAppend(
-                          { css: generated.css, processed: true },
-                          {
-                            css: removeWebpackTailwindGeneratedAssetCss(currentRawSource),
-                            processed: true,
-                          },
-                        )?.css ?? generated.css
-                        : generated.css,
-                      {
-                        cssPreflight: cssHandlerOptions.isMainChunk,
-                        generatedCss: true,
-                        preserveExistingPreflight: shouldPreserveExistingPreflight,
-                      },
-                    )
-                  : isWebGeneratorTarget
-                    ? finalizeCssAssetSource(generatorRawSource, { generatedCss: false })
-                    : finalizeCssAssetSource(
-                        (await compilerOptions.styleHandler(generatorRawSource, cssHandlerOptions)).css,
-                        { generatedCss: false },
-                      )
-                const sourceBareUserCss = isWebGeneratorTarget
-                  ? undefined
-                  : createWebpackGeneratorUserCssSourceAppend(
-                      ...[
-                        userRawSource,
-                        sourceCss === undefined
-                          ? undefined
-                          : {
-                              css: sourceCss,
-                              processed: sourceCssProcessed,
-                            },
-                        {
-                          css: generatorRawSource,
-                          processed: false,
-                        },
-                      ].map((source) => {
-                        if (source === undefined) {
-                          return undefined
-                        }
-                        return {
-                          css: collectWebpackBareSelectorUserCss(source.css),
-                          processed: source.processed,
-                        }
-                      }),
-                    )
-                const handledSourceBareUserCss = sourceBareUserCss === undefined
-                  ? ''
-                  : sourceBareUserCss.processed
-                    ? sourceBareUserCss.css
-                    : (await compilerOptions.styleHandler(sourceBareUserCss.css, cssHandlerOptions)).css
-                const finalizedSourceBareUserCss = handledSourceBareUserCss.trim().length === 0
-                  ? ''
-                  : finalizeCssAssetSource(handledSourceBareUserCss, {
-                      cssPreflight: false,
-                      generatedCss: false,
-                    })
-                const missingSourceBareUserCss = finalizedSourceBareUserCss.trim().length === 0
-                  ? ''
-                  : filterExistingCssRules(finalizedGeneratedCss, finalizedSourceBareUserCss)
-                const css = finalizeTracedCss(
-                  missingSourceBareUserCss.trim().length === 0
-                    ? finalizedGeneratedCss
-                    : createWebpackGeneratorUserCssSourceAppend(
-                      {
-                        css: finalizedGeneratedCss,
-                        processed: true,
-                      },
-                      {
-                        css: missingSourceBareUserCss,
-                        processed: true,
-                      },
-                    )!.css,
-                  cssHandlerOptions,
-                  { finalized: true },
-                )
-                const source = new ConcatSource(css)
-
-                if (generated) {
-                  for (const className of generated.classSet) {
-                    generatorRuntimeSet.add(className)
-                  }
-                  debug('css handle via tailwind v%s engine(%s): %s', runtimeState.tailwindRuntime.majorVersion, generated.target, file)
-                }
-                else {
-                  debug('css handle: %s', file)
-                }
-
-                return {
-                  result: source,
-                }
-              },
-            })
-          }, cssTaskFactories, 'tasks.css')
+          await processWebpackGeneratedCssAsset(element, cssTaskContext)
         }
-        if (!watchMode) {
-          pushConcurrentTaskFactories(tasks, htmlTaskFactories, taskConcurrency)
-          await Promise.all(tasks)
-          tasks.length = 0
-          pushConcurrentTaskFactories(tasks, jsTaskFactories, taskConcurrency)
-          await Promise.all(tasks)
-          tasks.length = 0
-          pushConcurrentTaskFactories(tasks, cssTaskFactories, taskConcurrency)
-        }
-
-        const taskWaitStartedAt = performance.now()
-        await Promise.all(tasks)
-        recordTimingDetail('tasks.wait', taskWaitStartedAt)
-        const pruneStartedAt = performance.now()
-        compilerOptions.cache.prune?.({
-          cacheKeys: activeProcessCacheKeys,
-          hashKeys: activeProcessHashKeys,
-        })
-        const activeCssFiles = new Set(groupedEntries.css.map(([file]) => file))
-        pruneWebpackCssHandlerOptionCaches(
+        await finalizeWebpackProcessAssets({
+          activeProcessCacheKeys,
+          activeProcessHashKeys,
+          activeWebpackCssSourceFiles,
+          compilerOptions,
+          cssAssetResources,
           cssHandlerOptionsCache,
+          cssTaskFactories,
           cssUserHandlerOptionsCache,
-          activeCssFiles,
-        )
-        if (activeCssFiles.size > 0) {
-          pruneWebpackCssSources?.(new Set([
-            ...registeredWebpackCssSourceFiles,
-            ...activeWebpackCssSourceFiles,
-            ...[...cssAssetResources.values()].flatMap(resources => [...resources]),
-          ]), { watchMode })
-        }
-        recordTimingDetail('cache.prune', pruneStartedAt)
-        debug('end')
-        emitHmrTiming('webpack', 'processAssets', performance.now() - hmrTimingStartedAt, {
-          ...timingDetails,
-          memoryDebug: resolveWebpackMemoryDebugStats({
-            activeAssetFiles: entries.length,
-            activeCssFiles: activeCssFiles.size,
-            activeProcessCacheKeys,
-            activeProcessHashKeys,
-            cache: compilerOptions.cache,
-            cssHandlerOptionsCache,
-            cssUserHandlerOptionsCache,
-            phase: 'processAssets',
-            sourceCandidateScan: webpackSourceCandidateScanCache.getMemoryStats(),
-          }),
+          debug,
+          entries,
+          groupedEntries,
+          hmrTimingStartedAt,
+          htmlTaskFactories,
+          jsTaskFactories,
+          pruneWebpackCssSources,
+          recordTimingDetail,
+          registeredWebpackCssSourceFiles,
+          taskConcurrency,
+          tasks,
+          timingDetails,
+          watchMode,
+          webpackSourceCandidateScanCache,
         })
-        compilerOptions.onEnd()
       },
     )
     if (!isWebGeneratorTarget) {
