+++ packages/weapp-tailwindcss/src/bundlers/webpack/BaseUnifiedPlugin/v5-assets/pipeline-helpers/runtime-candidates.ts
@@ -1,0 +1,7 @@
+export function isRuntimeTransformCandidate(candidate: string) {
+  return candidate.length > 0
+    && !candidate.includes('=')
+    && !candidate.includes('<')
+    && !candidate.includes('>')
+    && !candidate.includes('${')
+}
